Make suggestions translate on-demand
Previously, when suggesting completions, we would always translate it to all
locales at the moment it was requested for any locale.
Now, we only translate it to the requested locale and slowly accumulate the
localizations that way.

This could *still* cause suggestions to take a long time to show up.
(Discord actually has a very strict time-out as of now, about 5 seconds.)
But I don't think there is a much better option that doesn't completely blow
the time-budget for this project.